backend/requirements.txt
# FastAPI and server

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6


# OCR Engines

paddleocr==2.7.3
pytesseract==0.3.10


# Image Processing

opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.26.2


# Deep Learning (Windows-compatible CPU version)

paddlepaddle==2.6.1


# Database

sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9


# Task Queue

celery==5.3.4
redis==5.0.1


# Utilities

python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0


# File handling

PyPDF2==3.0.1
pdf2image==1.16.3
python-docx==1.2.0
reportlab==4.0.0

